+++ python2verilog/__init__.py
@@ -7,3 +7,5 @@ from . import backend
 from . import optimizer
 from . import utils
 from . import ir
+from . import api
+from .api import verilogify
