+++ python2verilog/api/__init__.py
@@ -2,11 +2,14 @@
 API wrappers
 """
 
-from .api import (
+from .cli import (
     parse_python,
     convert_for_debug,
     convert,
     convert_file_to_file,
+)
+
+from .decorator import (
     global_scope,
     verilogify,
 )
