+++ python2verilog/api/cli.py
@@ -26,148 +26,6 @@ from .. import ir
 from ..backend import verilog
 from ..optimizer import OptimizeGraph
 
-# All functions if a lesser namespace is not given
-global_scope: dict[FunctionType, ir.Context] = {}
-namespaces = [global_scope]
-
-
-def new_namespace():
-    """
-    Create new namespace and returns that namespace
-    """
-    namespace = {}
-    namespaces.append(namespace)
-    return namespace
-
-
-def __scope_exit_handler():
-    """
-    Handles the conversions in each namespace for program exit
-    """
-    for namespace in namespaces:
-        for _, value in namespace.items():
-            logging.info(
-                value.name, value.test_cases, value.input_types, value.output_types
-            )
-        for context in namespace.values():
-            ir_root, context = Generator2Graph(context, context.py_ast).results
-            if context.optimization_level > 0:
-                OptimizeGraph(ir_root, threshold=context.optimization_level - 1)
-            ver_code_gen = verilog.CodeGen(ir_root, context)
-            assert isinstance(context, ir.Context)
-
-            if context.write:
-                module_str = ver_code_gen.get_module_str()
-                tb_str = ver_code_gen.new_testbench_str(context.test_cases)
-                context.module_file.write(module_str)
-                context.testbench_file.write(tb_str)
-
-
-atexit.register(__scope_exit_handler)
-
-
-# pylint: disable=dangerous-default-value
-@decorator_with_args
-def verilogify(
-    func: FunctionType,
-    namespace: dict[FunctionType, ir.Context] = global_scope,
-    module_output: Optional[Union[os.PathLike, typing.IO]] = None,
-    testbench_output: Optional[Union[os.PathLike, typing.IO]] = None,
-    write: bool = False,
-    overwrite: bool = False,
-):
-    """
-    :param namespace: the namespace to put this function, for linking purposes
-    :param write: if True, files will be written to the specified paths
-    :param overwrite: If True, existing files will be overwritten
-    :param module_output: path to write verilog module to, defaults to function_name.sv
-    :param testbench_output: path to write verilog testbench to, defaults to function_name_tb.sv
-    """
-    if overwrite and not write:
-        raise RuntimeError("Overwrite is true, but write is set to false")
-    if func in namespace:
-        raise RuntimeError(f"{func.__name__} has already been decorated")
-
-    # Get caller filename for default output paths
-    # .stack()[2] as this function uses a decorator, so the first frames' filename
-    # is the filename that contains that decorator
-    filename = inspect.stack()[2].filename
-    input_file_stem = os.path.splitext(filename)[0]  # path with no extension
-
-    if not module_output:
-        module_output = Path(input_file_stem + ".sv")
-    if not testbench_output:
-        testbench_output = Path(input_file_stem + "_tb.sv")
-
-    tree = ast.parse(textwrap.dedent(inspect.getsource(func)))
-    assert len(tree.body) == 1
-    func_ast = tree.body[0]
-    assert isinstance(
-        func_ast, ast.FunctionDef
-    ), f"Got {type(func_ast)} expected {ast.FunctionDef}"
-
-    context = ir.Context(name=func.__name__)
-    context.py_ast = func_ast
-    context.py_func = func
-
-    context.input_vars = [ir.Var(name.arg) for name in func_ast.args.args]
-
-    context.write = write
-
-    if write:
-        if overwrite:
-            mode = "w"
-        else:
-            mode = "x"
-        # pylint: disable=consider-using-with
-        if isinstance(module_output, os.PathLike):
-            try:
-                module_output = open(module_output, mode=mode, encoding="utf8")
-            except FileExistsError as e:
-                raise FileExistsError("Try setting overwrite to True") from e
-        if isinstance(testbench_output, os.PathLike):
-            try:
-                testbench_output = open(testbench_output, mode=mode, encoding="utf8")
-            except FileExistsError as e:
-                raise FileExistsError("Try setting overwrite to True") from e
-        context.module_file = module_output
-        context.testbench_file = testbench_output
-
-    namespace[func] = context
-
-    @wraps(func)
-    def generator_wrapper(*args, **kwargs):
-        if kwargs:
-            raise RuntimeError(
-                "Keyword arguments not yet supported, use positional arguments only"
-            )
-        context = namespace[func]
-        context.test_cases.append(args)
-        if not context.input_types:
-            context.input_types = [type(arg) for arg in args]
-        else:
-            context.check_input_types(args)
-
-        for result in func(*args, **kwargs):
-            if not isinstance(result, tuple):
-                result = (result,)
-
-            if not context.output_types:
-                logging.info(f"Using {result} as reference")
-                context.output_types = [type(arg) for arg in result]
-            else:
-                logging.debug(f"Next yield gave {result}")
-                context.check_output_types(result)
-
-        return func(*args, **kwargs)
-
-    @wraps(func)
-    def function_wrapper(*args, **kwargs):
-        logging.error("Non-generator functions currently not supported")
-        return func(*args, **kwargs)
-
-    return generator_wrapper if inspect.isgeneratorfunction(func) else function_wrapper
-
 
 @overload
 def convert(context: str, code: str, optimization_level: int = 1):
